+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -73,6 +73,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Assign default volume profile when creating a new HDRP Asset
 - Fixed fov to 0 in planar probe breaking the projection matrix (case 1182014)
 - Fixed bugs with shadow caching 
+- Reassign the same camera for a realtime probe face render request to have appropriate history buffer during realtime probe rendering.
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled